src/requirement.txt
tensorflow-gpu
opencv-python
matplotlib
pandas
